requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
torch==2.2.1
transformers==4.40.1
accelerate==0.29.3
sentencepiece==0.2.0
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
requests==2.31.0
jieba==0.42.1
faiss-gpu==1.7.4
numpy==1.26.4
rank_bm25==0.2.2
scikit-learn==1.4.2
sentence-transformers==2.7.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
cn2an==0.5.22